requirements.txt
google-genai==1.66.0
beautifulsoup4==4.12.3
requests>=2.31.0
chromadb==1.5.5
sentence-transformers==5.3.0
Flask>=2.3.0
